Undo moving generateDataListItems to a hook
