+++ packages/components/src/DataList/__tests__/DataList.utils.test.tsx
@@ -1,14 +1,13 @@
 import { render, screen } from "@testing-library/react";
 import React from "react";
-import { renderHook } from "@testing-library/react-hooks";
 import {
   generateHeaderElements,
+  generateListItemElements,
   getCompoundComponent,
   getCompoundComponents,
   sortSizeProp,
 } from "../DataList.utils";
 import { BREAKPOINTS } from "../DataList.const";
-import { useListItemElements } from "../hooks/useListItemElements";
 
 const date = new Date("2001-01-01T00:00:00.000Z");
 
@@ -49,11 +48,9 @@ describe("Datalist utils", () => {
    * 1. If any of the snapshot here needs updating, it's either the code within
    *    the utils have changed or there really is something wrong.
    */
-  describe("useListItemElements", () => {
+  describe("generateListItemElements", () => {
     it("should generate a Text component for the label key", () => {
-      const elementList = renderHook(() =>
-        useListItemElements([{ id: 1, label: "Hello" }]),
-      ).result.current;
+      const elementList = generateListItemElements([{ id: 1, label: "Hello" }]);
 
       // Snapshot needs updating? See comment #1 above the `describe`.
       expect(elementList[0].label).toMatchInlineSnapshot(`
@@ -66,14 +63,12 @@ describe("Datalist utils", () => {
     });
 
     it("should generate a subdued Text component for any random key", () => {
-      const elementList = renderHook(() =>
-        useListItemElements([
-          {
-            id: 1,
-            randomKeyThatIsntNormal: "I am a normal text",
-          },
-        ]),
-      ).result.current;
+      const elementList = generateListItemElements([
+        {
+          id: 1,
+          randomKeyThatIsntNormal: "I am a normal text",
+        },
+      ]);
 
       // Snapshot needs updating? See comment #1 above the `describe`.
       expect(elementList[0].randomKeyThatIsntNormal).toMatchInlineSnapshot(`
@@ -86,9 +81,9 @@ describe("Datalist utils", () => {
     });
 
     it("should generate a list of inline label for the tag key", () => {
-      const elementList = renderHook(() =>
-        useListItemElements([{ id: 1, tags: ["uno", "dos"] }]),
-      ).result.current;
+      const elementList = generateListItemElements([
+        { id: 1, tags: ["uno", "dos"] },
+      ]);
 
       // Snapshot needs updating? See comment #1 above the `describe`.
       expect(elementList[0].tags).toMatchInlineSnapshot(`
@@ -104,18 +99,16 @@ describe("Datalist utils", () => {
     });
 
     it("should generate the element passed in on any key", () => {
-      const elementList = renderHook(() =>
-        useListItemElements([{ id: 1, element: <div /> }]),
-      ).result.current;
+      const elementList = generateListItemElements([
+        { id: 1, element: <div /> },
+      ]);
 
       // Snapshot needs updating? See comment #1 above the `describe`.
       expect(elementList[0].element).toMatchInlineSnapshot(`<div />`);
     });
 
     it("should generate the correct element", () => {
-      const elementList = renderHook(() =>
-        useListItemElements([{ id: 1, date }]),
-      ).result.current;
+      const elementList = generateListItemElements([{ id: 1, date }]);
 
       // Snapshot needs updating? See comment #1 above the `describe`.
       expect(elementList[0].date).toMatchInlineSnapshot(`
